Refactor monte_carlo: add core and config modules
Add dedicated monte_carlo/core.py and monte_carlo/config.py and reorganize monte_carlo package. MCPipeline was moved out of mc_constructs into core.py; simulation, data-generation, filter and wald test helper functions and MCReferenceConstruct were extracted into config.py. monte_carlo.__init__ updated to import/forward the new symbols. mc_constructs.py was trimmed to remove the duplicated pipeline and test helpers. Tests adjusted to use np.testing.assert_allclose for strict floating-point comparisons in Wald tests. This separates pipeline logic from configuration/data helpers and fixes fragile equality checks in tests.
